+++ test-resources/babashka/tasks_cli.clj
@@ -1,0 +1,7 @@
+(ns babashka.tasks-cli)
+
+(defn outdated [{:keys [opts]}]
+  (prn (assoc opts :ran :outdated)))
+
+(defn clean [{:keys [opts]}]
+  (prn (assoc opts :ran :clean)))